MPI enhance
fix bugs of transferring size_t variables into int type in gqtensor MPI send/recv functions.
introduce loop MPI send and recv to support transfer data with length large than int.
change some variable type of ushort to size_t in the class in ten_extra_ctrct.h.

+++ include/gqten/gqtensor/gqtensor_impl.h
@@ -629,7 +629,7 @@ void GQTensor<ElemT, QNT>::ConciseShow(const size_t indent_level) const {
     cout << IndentPrinter(indent_level + 1) << "default tensor" << endl;
     return;
   }
-  unsigned data_size = pblk_spar_data_ten_->GetActualRawDataSize();
+  size_t data_size = pblk_spar_data_ten_->GetActualRawDataSize();
   cout << IndentPrinter(indent_level + 1)
        << "actual data size:\t" << data_size << "\n";
   cout << IndentPrinter(indent_level + 1)
@@ -706,7 +706,7 @@ inline void send_gqten(boost::mpi::communicator world,
   int tag_data = tag * kMPIDataTagMultiplyFactor + 1;
   const BlockSparseDataTensor<ElemT, QNT> &bsdt = gqten.GetBlkSparDataTen();
   const ElemT *data_pointer = bsdt.GetActualRawDataPtr();
-  int data_size = bsdt.GetActualRawDataSize();
+  size_t data_size = bsdt.GetActualRawDataSize();
   ElemT zero = ElemT(0.0);
   if (gqten.IsScalar() && data_size == 0) {
     data_pointer = &zero;
@@ -756,7 +756,7 @@ inline boost::mpi::status recv_gqten(boost::mpi::communicator world,
   int tag_data = tag * kMPIDataTagMultiplyFactor + 1;
   BlockSparseDataTensor<ElemT, QNT> &bsdt = gqten.GetBlkSparDataTen();
   ElemT *data_pointer = bsdt.pactual_raw_data_;
-  int data_size = bsdt.GetActualRawDataSize();
+  size_t data_size = bsdt.GetActualRawDataSize();
 #ifdef GQTEN_MPI_TIMING_MODE
   Timer mpi_recv_gqten_data_timer("mpi_recv_gqten_data: from rank "
                                       + std::to_string(source)
@@ -791,7 +791,7 @@ inline void SendBroadCastGQTensor(
 
   const BlockSparseDataTensor<ElemT, QNT> &bsdt = gqten.GetBlkSparDataTen();
   const ElemT *raw_data_pointer = bsdt.GetActualRawDataPtr();
-  int raw_data_size = bsdt.GetActualRawDataSize();
+  size_t raw_data_size = bsdt.GetActualRawDataSize();
   ElemT zero = ElemT(0.0);
   if (gqten.IsScalar() && raw_data_size == 0) {
     raw_data_pointer = &zero;
@@ -828,7 +828,7 @@ inline void RecvBroadCastGQTensor(
 
   BlockSparseDataTensor<ElemT, QNT> &bsdt = gqten.GetBlkSparDataTen();
   ElemT *raw_data_pointer = bsdt.pactual_raw_data_;
-  int raw_data_size = bsdt.GetActualRawDataSize();
+  size_t raw_data_size = bsdt.GetActualRawDataSize();
 #ifdef GQTEN_MPI_TIMING_MODE
   Timer mpi_bcast_recv_gqten_data_timer("mpi_bcast_send_gqten_data: root = "
                                             + std::to_string(root)
